restructured constants and util
