+++ src/download.py
@@ -1,7 +1,5 @@
-import requests
-
-import util
-import constants
+from orphan_detection import constants
+from orphan_detection import util
 
 
 def download_step(searched_domain: str) -> str:
@@ -9,9 +7,8 @@ def download_step(searched_domain: str) -> str:
     zipped_archive_file = constants.ZIPPED_ARCHIVE_NAME_TEMPLATE.format(DOMAIN=searched_domain, DATE=date)
 
     # retrieve data from web archiv
-    response = requests.get(constants.WEB_ARCHIV_BASE_URL.format(DOMAIN=searched_domain))
-    response.raise_for_status()
+    response_data = util.download_web_archive_data(domain=searched_domain)
 
     # zip results and save in archive directory
-    util.save_to_gzip_file(zipped_archive_file, response.text)
+    util.write_lines_to_file(zipped_archive_file, response_data, zipped_file=True)
     return zipped_archive_file
